refactor(chat): Move messages into individual widgets.
- This breakup helps make the code readable and easy to maintain.
